refactor: フォルダ周りのロジック
