+++ src/app/(multi-footer)/[username]/page.client.tsx
@@ -1,13 +1,11 @@
 "use client";
-import { useState } from "react";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { useRouter, useSearchParams } from "next/navigation";
 import { Box, useMediaQuery } from "@mui/material";
 import type { ReflectionsCount } from "@/src/api/reflections-count-api";
 import type { TagType } from "@/src/hooks/reflection-tag/useExtractTrueTags";
 import type { User } from "@prisma/client";
-import { folderAPI, type Folder } from "@/src/api/folder-api";
+import { type Folder } from "@/src/api/folder-api";
 import {
-  reflectionAPI,
   type RandomReflection,
   type Reflection,
   type ReflectionTagCountList
@@ -24,9 +22,9 @@ import UserProfileArea from "@/src/features/routes/reflection-list/profile/UserP
 import { SelectionHeader } from "@/src/features/routes/reflection-list/selection-header/SelectionHeader";
 import { Sidebar } from "@/src/features/routes/reflection-list/sidebar";
 import { FolderInitializer } from "@/src/features/routes/reflection-list/sidebar/FolderInitializer";
+import { useFolderSelection } from "@/src/hooks/folder/useFolderSelection";
 import { usePagination } from "@/src/hooks/reflection/usePagination";
 import { tagMap } from "@/src/hooks/reflection-tag/useExtractTrueTags";
-import { getReflectionWithFolderInfo } from "@/src/utils/actions/get-reflection-with-folder-info";
 import { useFolderStore } from "@/src/utils/store/useFolderStore";
 
 type UserReflectionListPageProps = {
@@ -58,70 +56,21 @@ const UserReflectionListPage: React.FC<UserReflectionListPageProps> = ({
   randomReflection,
   folders
 }) => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [isSelectMode, setIsSelectMode] = useState(false);
-  const [selectedReflections, setSelectedReflections] = useState<string[]>([]);
   const router = useRouter();
-  const pathname = usePathname();
   const searchParams = useSearchParams();
   const isPWA = useMediaQuery("(display-mode: standalone)");
-
   const { handlePageChange } = usePagination();
 
-  const foldersState = useFolderStore((state) => state.folders);
-  const selectedInfo = useFolderStore((state) => state.selectedInfo);
-  const setFolders = useFolderStore((state) => state.setFolders);
-  const setSelectedInfo = useFolderStore((state) => state.setSelectedInfo);
-  const handleSelectMode = async (folderUUID: string) => {
-    router.push(pathname);
-    setIsSelectMode(true);
-    setSelectedInfo(folderUUID);
-    const info = await getReflectionWithFolderInfo(username, folderUUID);
-    if (!info) return;
-    const preSelected = info.map((i) => i.reflectionCUID);
-    setSelectedReflections(preSelected);
-  };
-
-  const handleSelect = (reflectionCUID: string) => {
-    setSelectedReflections((prev) => {
-      if (prev.includes(reflectionCUID)) {
-        return prev.filter((id) => id !== reflectionCUID);
-      } else {
-        return [...prev, reflectionCUID];
-      }
-    });
-  };
-
-  const handleCancelSelectMode = () => {
-    setIsSelectMode(false);
-    setSelectedReflections([]);
-    setSelectedInfo("");
-  };
-
-  const handleAddReflectionToFolder = async () => {
-    setIsLoading(true);
-
-    await reflectionAPI.bulkUpdateFolderReflection({
-      reflectionCUID: selectedReflections,
-      folderUUID: selectedInfo,
-      username
-    });
-    const updatedFolders = await folderAPI.getFolder(username);
-    if (updatedFolders === 404) {
-      return router.refresh();
-    }
-
-    setFolders(updatedFolders);
-    setIsSelectMode(false);
-    setSelectedReflections([]);
-    setIsLoading(false);
-    router.push(`/${username}?folder=${selectedInfo}`);
-    router.refresh();
-  };
-  const isFolderSelected = foldersState.some(
-    (f) => f.folderUUID === selectedInfo
-  );
-  const disableAdd = selectedReflections.length === 0 || isLoading;
+  const {
+    isSelectMode,
+    selectedReflections,
+    isFolderSelected,
+    disableAdd,
+    handleSelectMode,
+    handleSelect,
+    handleCancelSelectMode,
+    handleAddReflectionToFolder
+  } = useFolderSelection(username);
 
   const isCurrentUser = currentUsername === username;
   const isModalOpen = searchParams.get("status") === "posted";
@@ -129,6 +78,8 @@ const UserReflectionListPage: React.FC<UserReflectionListPageProps> = ({
     router.push(`/${username}`);
   };
 
+  const foldersState = useFolderStore((state) => state.folders);
+  const selectedInfo = useFolderStore((state) => state.selectedInfo);
   const currentFolder = foldersState.find((f) => f.folderUUID === selectedInfo);
   const selected = (() => {
     if (currentFolder) {
